audio_file_manager: split filesystem-agnostic code to separate file
